Remoção do pequenos warnings.
git-svn-id: http://svn.infox.com.br/repos/infoxnovoepa/aplicacao/trunk/epa2@389 069f99d3-da65-4ab0-85f1-cf03781be9e7

+++ src/main/java/br/com/infox/ibpm/home/PainelUsuarioHome.java
@@ -87,7 +87,7 @@ public class PainelUsuarioHome implements Serializable {
 	}
 	
 	public void processoCaixa(DropEvent evt) {
-		Caixa caixa = EntityUtil.find(Caixa.class, (Integer) evt.getDropValue());
+		Caixa caixa = EntityUtil.find(Caixa.class, evt.getDropValue());
 		setProcessoCaixa(getProcessoIdList(evt.getDragValue()), caixa);
 		try {
 			Processo cpt = (Processo) evt.getDragValue();
